Dealing with extra status codes

+++ lib/gig/downloader.rb
@@ -5,10 +5,13 @@ module Gig
       Dir.mkdir(directory) unless Dir.exist?(directory)
 
       path = "#{Dir.pwd}/#{directory}/#{filename}.jpg"
+      return false if File.exist?(path)
+
       URI.open(path, 'wb') do |file|
         puts "Downloading #{filename}.jpg to #{Dir.pwd}/#{directory}"
         file << URI.open(uri).read
       end
+      true
     end
   end
 end
